Revert "Use first 3 version bits to signal block type"
This reverts commit aec2364b3fe87f23f03bf3f3baaee830a9c5c1e6.

+++ src/rpc/mining.cpp
@@ -717,7 +717,7 @@ static UniValue getblocktemplate(const JSONRPCRequest& request)
             }
         }
     }
-    result.pushKV("version", (uint64_t)pblock->nVersion);
+    result.pushKV("version", pblock->nVersion);
     result.pushKV("rules", aRules);
     result.pushKV("vbavailable", vbavailable);
     result.pushKV("vbrequired", int(0));
